requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.7
    # via litellm
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via litellm
distro==1.9.0
    # via openai
filelock==3.18.0
    # via huggingface-hub
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.5.1
    # via huggingface-hub
h11==0.16.0
    # via httpcore
hf-xet==1.1.2
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langsmith
    #   litellm
    #   openai
huggingface-hub==0.32.3
    # via tokenizers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via litellm
jinja2==3.1.6
    # via litellm
jiter==0.10.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.24.0
    # via litellm
jsonschema-specifications==2025.4.1
    # via jsonschema
langchain==0.3.25
    # via -r requirements.in
langchain-core==0.3.63
    # via
    #   langchain
    #   langchain-text-splitters
langchain-text-splitters==0.3.8
    # via langchain
langsmith==0.3.43
    # via
    #   langchain
    #   langchain-core
litellm==1.72.0
    # via -r requirements.in
markupsafe==3.0.2
    # via jinja2
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
openai==1.82.1
    # via litellm
orjson==3.10.18
    # via langsmith
packaging==24.2
    # via
    #   huggingface-hub
    #   langchain-core
    #   langsmith
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
pydantic==2.11.5
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-core
    #   langsmith
    #   litellm
    #   openai
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.0
    # via litellm
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-core
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   huggingface-hub
    #   langchain
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.41
    # via langchain
tenacity==9.1.2
    # via langchain-core
tiktoken==0.9.0
    # via litellm
tokenizers==0.21.1
    # via litellm
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
typing-extensions==4.14.0
    # via
    #   anyio
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.4.0
    # via requests
yarl==1.20.0
    # via aiohttp
zipp==3.22.0
    # via importlib-metadata
zstandard==0.23.0
    # via langsmith
